Добавляет функцию форматирования продолжительности фильма
