+++ src/view/film-card.js
@@ -1,5 +1,6 @@
-import AbstractView from "./abstract";
 import dayjs from "dayjs";
+import AbstractView from "./abstract";
+import {formatFilmDuration} from "../utils/film";
 
 const createFilmCardTemplate = (film) => {
   const {poster, title, rating, release, duration, genres, description, comments: commentsIds, isWatch, isWatched, isFavorite} = film;
@@ -21,7 +22,7 @@ const createFilmCardTemplate = (film) => {
     <p class="film-card__rating">${rating}</p>
     <p class="film-card__info">
       <span class="film-card__year">${dayjs(release).format(`YYYY`)}</span>
-      <span class="film-card__duration">${duration}</span>
+      <span class="film-card__duration">${formatFilmDuration(duration)}</span>
       <span class="film-card__genre">${genres[0]}</span>
     </p>
     <img src="${poster}" alt="" class="film-card__poster">
